Setup project files for initial commit
This commit contains everything needed in order for this project to
work. For now though it only contains an example in the test file to
understand how the FFI between the two languages. This will be further
expanded to actually provide and easy to use overhead for the end user.