Implement FFI primitive types for both langs
This adds types to both Haskell and Rust that use the same naming
terminology between the two. This makes it easy to map types between the
two. In reality though they are aliases for the C lang types that they
represent underneath. These are more readable then their C equivalents.
It also exports the necessary modules needed so that the types actually
work when the libraries are imported.